break up solver types into own modules. update tests. update tutorials (working) and other notebooks
