+++ src/energy/energy.jl
@@ -1,11 +1,5 @@
 # Contains the energy module, for everything relating to energy transport
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
 module energy
 
     # System libraries
