+++ src/energy/rfm.jl
@@ -1,11 +1,5 @@
 # Contains module and functions for performing radiative transfer with RFM
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
 module rfm
 
     # Import modules
