+++ src/interface/paths.jl
@@ -1,5 +1,67 @@
 module paths
 
+    # Load modules
+    using LoggingExtras
+
+    # AGNI root directory (constant)
     const ROOT_DIR::String = abspath(dirname(abspath(@__FILE__)), "..", "..")
+    export ROOT_DIR
+
+    # Resources directory (constant)
+    const RES_DIR::String = joinpath(ROOT_DIR, "res")
+    export RES_DIR
+
+    # FWL_DATA folder (fall back to RES_DIR if not set)
+    const FWL_DATA::String = joinpath(get(ENV, "FWL_DATA", RES_DIR))
+    export FWL_DATA
+
+    # Get path to other data dirs (can be overridden)
+    function get_dir(name::String)::String
+
+        if name == "thermodynamics"
+            return joinpath(RES_DIR, "thermodynamics")
+
+        elseif name == "scattering"
+            return joinpath(RES_DIR, "scattering")
+
+        elseif name == "config"
+            return joinpath(RES_DIR, "config")
+
+        elseif name == "stellar_spectra"
+            return joinpath(RES_DIR, "stellar_spectra")
+
+        elseif name == "spectral_files"
+            return joinpath(RES_DIR, "spectral_files")
+
+        elseif name == "out"
+            return joinpath(ROOT_DIR, "out")
+
+        else
+            @error "Unknown directory name: $name"
+            return nothing
+        end
+    end
+    export get_dir
+
+    """
+    **Check if directory is 'safe' for removal**
+
+    Arguments:
+    - path::String                  the path to check
+
+    Returns:
+    - Bool                          true if the path is safe for removal
+    """
+    function is_safe_dir(path::String)::Bool
+        return !(isempty(path) ||                    # empty path
+                 ispath(joinpath(path, ".git")) ||   # contains git repo
+                 (joinpath(path) == pwd()) ||        # is current working directory
+                 samefile(path, "/") ||              # is system root directory
+                 samefile(path, homedir()) ||        # is user home directory
+                 samefile(path, paths.ROOT_DIR) ||   # is AGNI root directory
+                 samefile(path, paths.RES_DIR)       # is AGNI resources directory
+                )
+    end
+    export is_safe_dir
 
 end
