+++ src/interface/save.jl
@@ -1,12 +1,5 @@
 # Write atmosphere to NetCDF or CSV file
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
-
 module save
 
     import ..atmosphere
