+++ src/phys/phys.jl
@@ -1,11 +1,5 @@
 # Contains physical data
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
 module phys
 
     # Import packages
