+++ src/state/atmosphere.jl
@@ -1,18 +1,10 @@
 # Contains the atmosphere module
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
 """
 **Main atmosphere module for storing model data**
 
 This module primarily contains the `Atmos_t` struct which does most of the heavy lifting
 in AGNI. There are also functions for setting up and configuring the struct.
-
-Also includes hydrostatic integrator and ways to estimate observable quantities.
 """
 module atmosphere
 
@@ -32,18 +24,7 @@ module atmosphere
     import ..phys
     import ..spectrum
     import ..paths
-
-    # Code versions
-    const AGNI_VERSION::String     = "1.10.0"  # current agni version
-    const SOCVER_minimum::Float64  = 2407.2    # minimum required socrates version
-
-    # Hydrostatic+gravity+mass calculation (constants and limits)
-    HYDROGRAV_steps::Int64   = 2000      # total number of steps in height integration
-    HYDROGRAV_maxdr::Float64 = 1e8       # maximum dz across each layer [m]
-    HYDROGRAV_mindr::Float64 = 1e-5      # minimum dz across each layer [m]
-    HYDROGRAV_ming::Float64  = 1e-4      # minimum allowed gravity [m/s^2]
-    HYDROGRAV_constg::Bool   = false     # constant gravity with height?
-    HYDROGRAV_selfg::Bool    = true      # include self-gravity of the atmosphere?
+    import ..consts: UNSET_STR, AGNI_VERSION, SOCVER_minimum
 
     # Configuration defaults
     const CFG_surface_material::String  = "greybody"
@@ -99,7 +80,6 @@ module atmosphere
     const CFG_ocean_cs_height::Float64      = 3000.0
 
     # Variable limits and defaults
-    const UNSET_STR::String             = "__AGNI_UNSET_STR"
     const NLEV_minimum::Int64           = 15        # minimum allowed number of levels
     const PHS_TIMESCALE_MIN::Float64    = 0.01      # minimum phase change timescale [s]
     const SURF_ROUGHNESS_MIN::Float64   = 1e-5      # [m]
@@ -629,18 +609,14 @@ module atmosphere
         end
 
         # Locate data directories
-        atmos.THERMO_DIR = joinpath(atmos.ROOT_DIR, "res", "thermodynamics")
-        atmos.SCATTERING_DIR = joinpath(atmos.ROOT_DIR, "res", "scattering")
+        atmos.THERMO_DIR = paths.get_dir("thermodynamics")
+        atmos.SCATTERING_DIR = paths.get_dir("scattering")
 
         # Make output directory if does not exist
-        if isempty(OUT_DIR)
-            @error "Output directory cannot be an empty path"
-            return false
-        end
         atmos.OUT_DIR = abspath(OUT_DIR)
-        if samefile(atmos.OUT_DIR, atmos.ROOT_DIR)
-            @error "Output directory cannot be the AGNI root directory"
-            @error "    Got: $(atmos.OUT_DIR)"
+        if !paths.is_safe_dir(atmos.OUT_DIR)
+            @error "Output directory is not safe to write to!"
+            @error "    OUT_DIR=$(atmos.OUT_DIR)"
             return false
         end
         if !isdir(atmos.OUT_DIR) && !isfile(atmos.OUT_DIR)
@@ -1447,510 +1423,6 @@ module atmosphere
 
 
     """
-    **Estimate photosphere.**
-
-    Estimates the location of the photosphere by finding the median of the contribution
-    function in each band (0.2 um to 150 um), and then finding the pressure level at which
-    these median values are maximised.
-
-    Arguments:
-        - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-
-    Returns:
-        - `p_ref::Float64`          pressure level of photosphere [Pa]
-    """
-    function estimate_photosphere!(atmos::atmosphere.Atmos_t)::Float64
-
-        # Params
-        wl_min::Float64  = 0.2 * 1e-6 # 200 nm
-        wl_max::Float64  = 150 * 1e-6 # 150 um
-        p_min::Float64   = 10.0 # 1e-4 bar
-
-        # tracking
-        cff_max::Float64 = 0.0
-        cff_try::Float64 = 0.0
-        atmos.transspec_p = p_min
-
-        # get band indices
-        wl_imin = findmin(abs.(atmos.bands_cen .- wl_min))[2]
-        wl_imax = findmin(abs.(atmos.bands_cen .- wl_max))[2]
-
-        # reversed?
-        if wl_imin > wl_imax
-            wl_imin, wl_imax = wl_imax, wl_imin
-        end
-
-        # loop over levels
-        for i in 1:atmos.nlev_c
-            if atmos.p[i] < p_min
-                continue
-            end
-
-            # maximum contfunc in this band
-            cff_try = Statistics.median(atmos.contfunc_band[i,wl_imin:wl_imax])
-
-            # is this more than the existing maximum?
-            if cff_try > cff_max
-                cff_max = cff_try
-                atmos.transspec_p = atmos.p[i]
-            end
-        end
-
-        return atmos.transspec_p
-    end
-
-    """
-    **Calculate observed radius and bulk density.**
-
-    This is done at the layer probed in transmission, which is set to a fixed pressure.
-
-    Arguments:
-        - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-
-    Returns:
-        - `transspec_rho::Float64`  the bulk density observed in transmission
-    """
-    function calc_observed_rho!(atmos::atmosphere.Atmos_t)::Float64
-
-        # transspec_r::Float64            # planet radius probed in transmission [m]
-        # transspec_m::Float64            # mass [kg] of atmosphere + interior
-        # transspec_rho::Float64          # bulk density [kg m-3] implied by r and m
-
-        # Store reference pressure in atmos struct
-        # estimate_photosphere!(atmos)
-
-        # get the observed height
-        idx::Int64 = findmin(abs.(atmos.p .- atmos.transspec_p))[2]
-        atmos.transspec_r    = atmos.r[idx]
-        atmos.transspec_μ    = atmos.layer_μ[idx]
-        atmos.transspec_tmp  = atmos.tmp[idx]
-        atmos.transspec_grav = atmos.g[idx]
-
-        # get mass of whole atmosphere, assuming hydrostatic
-        atmos.transspec_m = atmos.p_boa * 4 * pi * atmos.rp^2 / atmos.grav_surf
-
-        # add mass of the interior component
-        atmos.transspec_m += atmos.interior_mass
-
-        # get density of all enclosed by observed layer
-        # store this in the atmosphere struct
-        atmos.transspec_rho = 3.0 * atmos.transspec_m / (4.0 * pi * atmos.transspec_r^3)
-
-        # also return the value
-        return atmos.transspec_rho
-    end
-
-    """
-    **Get pressure at top and bottom of convective zone**
-
-    Arguments:
-        - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-
-    Returns:
-        - `p_top::Float64`          pressure [Pa] at top of convective zone
-        - `p_bot::Float64`          pressure [Pa] at bottom of convective zone
-    """
-    function estimate_convective_zone(atmos::atmosphere.Atmos_t)::Tuple{Float64,Float64}
-
-        # Defaults to zero, if there's no convection
-        p_top::Float64 = 0.0
-        p_bot::Float64 = 0.0
-
-        # Loop from top-down to find p_top
-        for i in 1:atmos.nlev_l
-            if atmos.mask_c[i]
-                p_top = atmos.pl[i]
-                break
-            end
-        end
-
-        # Loop from bottom-up to find p_bot
-        for i in range(start=atmos.nlev_l, stop=1, step=-1)
-            if atmos.mask_c[i]
-                p_bot = atmos.pl[i]
-                break
-            end
-        end
-
-        # Return top, bot
-        return (p_top, p_bot)
-    end
-
-    """
-    **Calculate properties within each layer of the atmosphere (e.g. density, mmw).**
-
-    Function will return false if hydrostatic calculcation fails. This is usually when
-    the atmosphere becomes unbound.
-
-    Arguments:
-        - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-
-    Returns:
-        - `ok::Bool`                function result is ok
-    """
-    function calc_layer_props!(atmos::atmosphere.Atmos_t)::Bool
-        if !atmos.is_param
-            @error("Atmosphere struct has not been setup")
-            return false
-        end
-
-        # Status
-        ok::Bool = true
-
-        # MMW
-        calc_profile_mmw!(atmos)
-
-        # Heat capacity and thermal conductivity
-        calc_profile_cpkc!(atmos)
-
-        # Set density at each level
-        calc_profile_density!(atmos)
-
-        # Perform hydrostatic integration
-        ok = ok && calc_profile_radius!(atmos)
-
-        # Calculate scale height at each layer
-        calc_profile_Hp!(atmos)
-
-        return ok
-    end
-
-    """
-    **Calculate radii, gravities, and masses for all layers.**
-
-    Performs hydrostatic integration from the ground upwards.
-    Requires density, temperature, pressure to have already been set.
-
-    Does not account for surface ocean height.
-
-    Arguments:
-    - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-
-    Returns:
-    - `bound::Bool`             atmosphere is strongly bound by gravity
-    """
-    function calc_profile_radius!(atmos::atmosphere.Atmos_t)::Bool
-
-        # Calculate net surface acceleration [m s-2]
-        a_surf::Float64 = atmos.grav_surf -
-                            phys.cent_accel(atmos.axial_period, atmos.rp, atmos.col_lat)
-
-        # Reset arrays
-        fill!(atmos.r         ,    atmos.rp)
-        fill!(atmos.rl        ,    atmos.rp)
-        fill!(atmos.g         ,    atmos.grav_surf)
-        fill!(atmos.gl        ,    atmos.grav_surf)
-        fill!(atmos.a         ,    a_surf)
-        fill!(atmos.al        ,    a_surf)
-        fill!(atmos.m         ,    atmos.interior_mass)
-        fill!(atmos.ml        ,    atmos.interior_mass)
-        fill!(atmos.layer_thick,   1.0)
-        fill!(atmos.layer_σ ,      1.0)
-        fill!(atmos.layer_isbound, true)
-
-        # Check config...
-        if HYDROGRAV_constg && HYDROGRAV_selfg
-            @warn "Incompatible gravity parameters have been set:"
-            @warn "    constant with height (HYDROGRAV_constg=$HYDROGRAV_constg)"
-            @warn "    atmos self-attraction (HYDROGRAV_selfg=$HYDROGRAV_selfg)"
-        end
-
-        # Temporary values
-        nsub::Int64 = round(Int64, HYDROGRAV_steps/atmos.nlev_c, RoundUp)
-
-        # Integrate from surface upwards
-        for i in range(start=atmos.nlev_c, stop=1, step=-1)
-
-            # ------------
-            # Integrate from lower edge to centre
-            atmos.r[i], atmos.g[i], atmos.m[i] =
-                integ_hydrograv(atmos.rl[i+1],
-                                    atmos.gl[i+1], atmos.al[i+1],
-                                    atmos.ml[i+1], atmos.pl[i+1],
-                                    atmos.p[i], atmos.layer_ρ[i], nsub)
-
-            #   apply radius limiter
-            atmos.r[i] = max(atmos.r[i], atmos.rl[i+1] + HYDROGRAV_mindr)
-            if atmos.r[i] > atmos.rl[i+1] + HYDROGRAV_maxdr/2
-                atmos.r[i] = atmos.rl[i+1] + HYDROGRAV_maxdr/2
-                atmos.layer_isbound[i] = false
-            end
-
-            #   apply gravity limiter
-            if HYDROGRAV_constg
-                atmos.g[i]  = atmos.grav_surf
-            end
-            if atmos.g[i] < HYDROGRAV_ming
-                atmos.g[i] = HYDROGRAV_ming
-                atmos.layer_isbound[i] = false
-            end
-
-            # calculate net acceleration at layer centre
-            atmos.a[i] = atmos.g[i] -
-                            phys.cent_accel(atmos.axial_period, atmos.r[i], atmos.col_lat)
-
-            # ------------
-            # Integrate from centre to upper edge
-            atmos.rl[i], atmos.gl[i], atmos.ml[i] =
-                integ_hydrograv(atmos.r[i],
-                                    atmos.g[i], atmos.a[i],
-                                    atmos.m[i], atmos.p[i],
-                                    atmos.pl[i], atmos.layer_ρ[i], nsub)
-
-            #   apply radius limiter
-            atmos.rl[i] = max(atmos.rl[i], atmos.r[i] + HYDROGRAV_mindr)
-            if atmos.rl[i] > atmos.r[i] + HYDROGRAV_maxdr/2
-                atmos.rl[i] = atmos.r[i] + HYDROGRAV_maxdr/2
-                atmos.layer_isbound[i] = false
-            end
-
-            #   apply gravity limiter
-            if HYDROGRAV_constg
-                atmos.gl[i] = atmos.grav_surf
-            end
-            if atmos.gl[i] < HYDROGRAV_ming
-                atmos.gl[i] = HYDROGRAV_ming
-                atmos.layer_isbound[i] = false
-            end
-
-            #  calculate net acceleration at layer upper edge
-            atmos.al[i] = atmos.gl[i] -
-                            phys.cent_accel(atmos.axial_period, atmos.rl[i], atmos.col_lat)
-
-            # Store: Layer geometrical thickness [m]
-            atmos.layer_thick[i] = atmos.rl[i] - atmos.rl[i+1]
-
-            # Mass of layer, per unit area at layer-centre [kg m-2]
-            atmos.layer_σ[i] = (atmos.ml[i] - atmos.ml[i+1])/(4 * pi * atmos.r[i]^2)
-        end
-
-        return all(atmos.layer_isbound)
-    end
-
-    """
-    **Integrate hydrostatic and gravity equations across a pressure interval.**
-
-    Uses the classic fourth-order Runge-Kutta method.
-
-    Internally, this function integrates from p0 to p1, which means that the pressure is
-    decreasing across the interval. The gravity is updated at each integration step,
-    and is offset by the centrifugal acceleration.
-
-    Arguments:
-    - `r0::Float64`     radius   at start of interval [m]
-    - `g0::Float64`     gravity  at start of interval [m s-2]
-    - `a0::Float64`     net accel at start of interval (grav - cent) [m s-2]
-    - `m0::Float64`     mass enc at start of interval [kg]
-    - `p0::Float64`     pressure at start of interval [Pa]
-    - `p1::Float64`     pressure at end   of interval [Pa]
-    - `rho::Float64`    density throughout interval, constant [kg m-3]
-    - `n::Int64`        number of steps for integration (n >= 2)
-
-    Returns:
-    - `rj::Float64`     radius   at end of interval [m]
-    - `gj::Float64`     gravity  at end of interval [kg]
-    - `mj::Float64`     mass enc at end of interval [kg]
-    """
-    function integ_hydrograv(r0::Float64, g0::Float64, a0::Float64, m0::Float64, p0::Float64,
-                                    p1::Float64, rho::Float64, n::Int64)::Tuple{Float64,Float64,Float64}
-
-        # Work variables
-        pj::Float64 = p0    # rolling pressure (decreasing)
-        rj::Float64 = r0    # rolling radius   (increasing)
-        gj::Float64 = g0    # rolling gravity  (incr, decr, or constant)
-        mj::Float64 = m0    # rolling mass     (increasing)
-
-        # Get centripetal acceleration's rotation factor, (2 pi cosθ / period)^2
-        # Backing-out the factor this way avoids extra trig function calls.
-        cent_fact::Float64 = (g0 - a0)/r0  # units: s-2
-
-        # Get gravitational acceleration at r
-        function _grav(r)
-            if HYDROGRAV_constg
-                # gravity is constant
-                return g0
-            else
-                if HYDROGRAV_selfg
-                    # gravity changes with mass and radius
-                    return phys.grav_accel(mj, rj)
-                else
-                    # gravity changes with radius only
-                    return g0 * (r0/rj)^2
-                end
-            end
-        end
-
-        # Get net acceleration at r
-        function _accel(r)
-            return _grav(r) - r*cent_fact
-        end
-
-        # Derivative to integrate
-        #   dr/dp = -1 / (rho * a(r))
-        function _drdp(p,r)
-            return -1 / (rho * _accel(r))
-        end
-
-        # Parameters
-        dp::Float64  = (p1-p0)/max(2,n) # this will be negative
-        dp2::Float64 = dp/2
-        k1::Float64  = 0.0; k2::Float64 = 0.0
-        k3::Float64  = 0.0; k4::Float64 = 0.0
-
-        # Loop over sub-levels between p0 and p1
-        for _ in range(p0, stop=p1, step=dp)
-
-            # Integrate radius ...
-            k1 = _drdp(pj,       rj)
-            k2 = _drdp(pj + dp2, rj + k1*dp2)
-            k3 = _drdp(pj + dp2, rj + k2*dp2)
-            k4 = _drdp(pj + dp,  rj + k3*dp)
-            rj += dp/6 * (k1 + 2*k2 + 2*k3 + k4)
-
-            # Integrate mass enclosed ...
-            mj += 4 * pi * rj^2 * (-1 * dp) / gj
-
-            # Integrate pressure (negative change )
-            pj += dp
-
-            # Update gravity
-            gj = _grav(rj)
-        end
-
-        return (rj, gj, mj)
-    end
-
-    """
-    **Calculate mean molecular weight for all layers.**
-
-    MMW stored as kg/mol.
-
-    Arguments:
-        - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function calc_profile_mmw!(atmos::atmosphere.Atmos_t)
-        fill!(atmos.layer_μ, 0.0)
-        for gas in atmos.gas_names
-            @. atmos.layer_μ += atmos.gas_vmr[gas] * atmos.gas_dat[gas].mmw
-        end
-        return nothing
-    end
-
-    """
-    **Calculate specific heat capacity and thermal conductivity for all layers.**
-
-    Specific heat per unit mass: J K-1 kg-1.
-    Thermal conductivity: W m-1 K-1.
-
-    Arguments:
-        - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function calc_profile_cpkc!(atmos::atmosphere.Atmos_t)
-        # Loop over layers
-        @inbounds for i in 1:atmos.nlev_c
-            calc_single_cpkc!(atmos, i)
-        end
-        return nothing
-    end
-
-    """
-    **Calculate specific heat capacity and thermal conductivity of a single layer.**
-
-    Specific heat per unit mass: J K-1 kg-1.
-    Thermal conductivity: W m-1 K-1.
-
-    Arguments:
-        - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-        - `idx::Int64`          index of the layer
-    """
-    function calc_single_cpkc!(atmos::atmosphere.Atmos_t, idx::Int64)
-        # Reset
-        mmr::Float64 = 0.0
-        atmos.layer_cp[idx] = 0.0
-        atmos.layer_kc[idx] = 0.0
-        # Loop over gases
-        for gas in atmos.gas_names
-            mmr = atmos.gas_vmr[gas][idx] * atmos.gas_dat[gas].mmw/atmos.layer_μ[idx]
-            atmos.layer_cp[idx] += mmr * phys.get_Cp(atmos.gas_dat[gas], atmos.tmp[idx])
-            atmos.layer_kc[idx] += mmr * phys.get_Kc(atmos.gas_dat[gas], atmos.tmp[idx])
-        end
-        return nothing
-    end
-
-    """
-    **Calculate the mass-density for all layers.**
-
-    Requires temperature, pressure, mmw to be already have been set.
-
-    Arguments:
-        - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function calc_profile_density!(atmos::atmosphere.Atmos_t)
-        # Loop over levels
-        @inbounds for i in 1:atmos.nlev_c
-            calc_single_density!(atmos, i)
-        end
-        return nothing
-    end
-
-    """
-    **Calculate the mass-density of a single layer.**
-
-    Requires temperature, pressure, mmw to be already have been set.
-
-    Arguments:
-    - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    - `idx::Int64`          index of the layer
-    """
-    function calc_single_density!(atmos::atmosphere.Atmos_t, idx::Int64)
-        # Array of gas properties
-        gas_arr = [atmos.gas_dat[gas]      for gas in atmos.gas_names]
-        vmr_arr = [atmos.gas_vmr[gas][idx] for gas in atmos.gas_names]
-
-        # Evaluate density
-        atmos.layer_ρ[idx] = phys.calc_rho_mix(
-                                                    gas_arr, vmr_arr,
-                                                    atmos.tmp[idx], atmos.p[idx],
-                                                    atmos.layer_μ[idx]
-                                                    )
-
-        return nothing
-    end
-
-        """
-    **Calculate the scale height for all layers.**
-
-    Requires temperature, pressure, mmw to be already have been set.
-
-    Arguments:
-        - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function calc_profile_Hp!(atmos::atmosphere.Atmos_t)
-        @inbounds for i in 1:atmos.nlev_c
-            calc_single_Hp!(atmos, i)
-        end
-        return nothing
-    end
-
-    """
-    **Calculate the scale height of a single layer.**
-
-    Requires temperature, pressure, mmw to be already have been set.
-
-    Arguments:
-    - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    - `idx::Int64`          index of the layer
-    """
-    function calc_single_Hp!(atmos::atmosphere.Atmos_t, idx::Int64)
-        if atmos.real_gas
-            atmos.layer_Hp[idx] = atmos.p[idx] / (atmos.layer_ρ[idx] * atmos.g[idx])
-        else
-            atmos.layer_Hp[idx] = phys.R_gas * atmos.tmp[idx] / (atmos.layer_μ[idx] * atmos.g[idx])
-        end
-        return nothing
-    end
-
-    """
     **Generate pressure grid.**
 
     Almost equally log-spaced between p_boa and p_boa. The near-boundary layers
@@ -2462,7 +1934,7 @@ module atmosphere
         #    hydrostatic integrator, this can cause dr to blow up, especially
         #    with a low MMW gas. Should be okay as long as the T(p) provided
         #    by the user is more reasonable.
-        calc_layer_props!(atmos)
+        # calc_layer_props!(atmos)
 
         ################################
         # Aerosol processes
@@ -2911,69 +2383,6 @@ module atmosphere
         return arr_P, arr_T, arr_R
     end
 
-    """
-    **Estimate a diagnostic Rayleigh number in each layer.**
-
-    Assuming that the Rayleigh number scales like `Ra ~ (wλ/κ)^(1/β)`
-    Where `κ` is the thermal diffusivity and `β` is the convective beta parameter.
-
-    This quantity must be taken lightly.
-
-    Arguments:
-    - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function estimate_Ra!(atmos::atmosphere.Atmos_t)
-
-        # Thermal diffusivity array
-        κ::Array{Float64,1} = zero(atmos.layer_cp)
-        @. κ = phys.calc_therm_diffus(atmos.layer_kc, atmos.layer_ρ, atmos.layer_cp)
-
-        # One over beta
-        ooβ::Float64 = 1.0 / phys.βRa
-
-        # Estimate Rayleigh number
-        @inbounds for i in 1:atmos.nlev_c
-            atmos.diagnostic_Ra[i] = ( atmos.w_conv[i] * atmos.λ_conv[i] / κ[i]) ^ ooβ
-        end
-
-        return nothing
-    end
-
-    """
-    **Estimate a diagnostic radiative timescale in each layer.**
-
-    This quantity must be taken lightly.
-
-    Arguments:
-    - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function estimate_timescale_rad!(atmos::atmosphere.Atmos_t)
-
-        # Equation 10.1 from Seager textbook
-        @inbounds for i in 1:atmos.nlev_c
-            atmos.timescale_rad[i] = atmos.layer_cp[i] * (atmos.pl[i+1] - atmos.pl[i]) /
-                                     (atmos.g[i] * 4 * phys.σSB * atmos.tmp[i]^3)
-        end
-
-        return nothing
-    end
-
-    """
-    **Estimate a diagnostic convective timescale in each layer.**
-
-    This quantity must be taken lightly.
-
-    Arguments:
-    - `atmos::Atmos_t`      the atmosphere struct instance to be used.
-    """
-    function estimate_timescale_conv!(atmos::atmosphere.Atmos_t)
-
-        @inbounds for i in 1:atmos.nlev_c
-            atmos.timescale_conv[i] = atmos.λ_conv[i] / max(atmos.w_conv[i], 1e-300)
-        end
-
-        return nothing
-    end
 
     """
     **List available aerosol species.**
