+++ src/state/multicol.jl
@@ -1,11 +1,5 @@
 # Contains the multicol module
 
-# Not for direct execution
-if (abspath(PROGRAM_FILE) == @__FILE__)
-    thisfile = @__FILE__
-    error("The file '$thisfile' is not for direct execution")
-end
-
 """
 **Main multicol module for handling multiple 1D-atmospheric columns**
